config: Use toml to store repo info. Bug fixes. PR feedback.
